I've added the SapphireUtils class to centralize the location of all the string variables, though I suppose I could move the to strings.xml instead. Maybe I'll do that through a refractor later.
I've also started hunting down bugs that persist between the Main UI, implementing the assistant, and implementing the TTS. I need them to be able to run generally bug free for the Alpha release. Centralizing all of the reference variables should help a lot since I won't have to manually change things anymore.

I'm also in the process of adding state variables to CoreService so that either the Main UI OR the system attempts to implement the TTS service and the STT service. Once these things are done it should be a generally more robust system.

That aside, for some reason, the CoreRegistrationService is giving an error when trying to send an intent back to CoreService. I hope it's not  because it's a background service (it may be though, since CoreService isn't a foreground service yet